Believe Part A is done here, I've just set it up so it takes into account the name of players in the player file. It takes in the config file and adds datastructure of linkedList (arrayList technically) in an array. To search for a person you'll search through the entire array until you find index beginning with P, then you've found where a persons attributes start.
I've put this into all three player files as the assignment is set up so
we can use different datastructures for all players if we want. If you
don't like the way it's set up you can change it in your task B or C.
Choose a task b or C and then you pretty much have responsibility of
either the BinaryGuessPlayer Class of the RandomGuessClass.

There's also Personal Attributes datastructure and total Attribute
datastructure.